Refactor #21: Move OasDiffResult to Services namespace for better organization
