+++ src/ConcordIO.Tool/Services/IOasDiffRunner.cs
@@ -1,5 +1,3 @@
-using ConcordIO.Tool.AOComparison;
-
 namespace ConcordIO.Tool.Services;
 
 /// <summary>
@@ -22,4 +20,15 @@ public interface IOasDiffRunner
     /// <param name="arguments">The command line arguments.</param>
     /// <returns>Result containing exit code, output, and error streams.</returns>
     Task<OasDiffResult> Run(string arguments);
+}
+
+/// <summary>
+/// Result of an oasdiff command execution.
+/// </summary>
+public class OasDiffResult
+{
+    public int ExitCode { get; init; }
+    public string Output { get; init; } = string.Empty;
+    public string Error { get; init; } = string.Empty;
+    public bool Breaking { get; init; }
 }
